requirements.txt
# Google AI SDKs
google-genai>=1.0.0
google-generativeai>=0.7.0

# HTTP requests
requests>=2.28.0

# Type hints support (for older Python versions)
typing-extensions>=4.0.0; python_version < '3.8'